+++ u-sdk/CHANGELOG.md
@@ -9,10 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- [lib]
+[lib]
+
 - 添加`credentials`模块，定义`Credentials` trait和相关实现供阿里云各个sdk模块使用
 
-- [oss]
+[oss]
+
 - 添加`GetObject`预签名URL的方法
 - 添加`PutObject`预签名URL的方法
 - 添加`PostObject`获取签名信息的方法
@@ -22,23 +24,33 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - 添加oss callback服务器端验证的axum Layer
 - 为现有方法添加sts临时凭证支持
 
+[sts]
 
-- [sts]
 - 添加sts模块，实现`AssumeRole`功能
 - 添加构建policy相关的方法
 - 添加临时凭证的Credentials功能
 
+[translate]
 
-- [translate]
+- 支持sts临时凭证进行相关api调用
+
+[email]
+
 - 支持sts临时凭证进行相关api调用
 
 ### Changed
 
-- [oss]
+[oss]
+
 - 更新和完善了代码/测试/用户的文档说明
 - 重构/简化了签名模块的函数参数结构
 
+[email]
+
+- email sdk使用OpenAPI V3签名版本
+
 ### Fixed
 
-- [oss]
+[oss]
+
 - 修复了`GetObject`header和query没有区分的问题